Move Map to its own component
